requirements.txt
# LiangClaw - 阿亮老師 AI Agent 平台
# Core
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
httpx>=0.28.0

# GUI
gradio>=5.0.0

# Bot Interfaces
python-telegram-bot>=20.7
line-bot-sdk>=3.11.0

# AI Providers (google-genai already installed)
# google-genai>=1.0.0

# PDF Generation (reportlab already installed)
# reportlab>=4.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
